table card and prettier added
