--- frontend/app/src/components/TimeRelatedFormExamples.tsx
@@ -1,91 +1,0 @@
-import React from "react"
-import { Button, DatePicker, Form, TimePicker } from "antd"
-import { StyleConstants } from "../styles/styleConstants"
-
-const { RangePicker } = DatePicker
-
-const formItemLayout = {
-  labelCol: {
-    xs: { span: StyleConstants.XS_BREAKPOINT },
-    sm: { span: StyleConstants.SM_LABEL_COL },
-  },
-  wrapperCol: {
-    xs: { span: StyleConstants.XS_BREAKPOINT },
-    sm: { span: StyleConstants.SM_WRAPPER_COL },
-  },
-}
-
-const config = {
-  rules: [
-    { type: "object" as const, required: true, message: "Please select time!" },
-  ],
-}
-
-const rangeConfig = {
-  rules: [
-    { type: "array" as const, required: true, message: "Please select time!" },
-  ],
-}
-
-const onFinish = (fieldsValue: any) => {
-  // Should format date value before submit.
-  const rangeValue = fieldsValue["range-picker"]
-  const rangeTimeValue = fieldsValue["range-time-picker"]
-  const values = {
-    ...fieldsValue,
-    "date-picker": fieldsValue["date-picker"].format("YYYY-MM-DD"),
-    "date-time-picker": fieldsValue["date-time-picker"].format(
-      "YYYY-MM-DD HH:mm:ss"
-    ),
-    "month-picker": fieldsValue["month-picker"].format("YYYY-MM"),
-    "range-picker": [
-      rangeValue[0].format("YYYY-MM-DD"),
-      rangeValue[1].format("YYYY-MM-DD"),
-    ],
-    "range-time-picker": [
-      rangeTimeValue[0].format("YYYY-MM-DD HH:mm:ss"),
-      rangeTimeValue[1].format("YYYY-MM-DD HH:mm:ss"),
-    ],
-    "time-picker": fieldsValue["time-picker"].format("HH:mm:ss"),
-  }
-  console.log("Received values of form: ", values)
-}
-
-const TimeRelatedFormExample: React.FC = () => (
-  <Form
-    name="time_related_controls"
-    {...formItemLayout}
-    onFinish={onFinish}
-    style={{ maxWidth: 600, margin: "0 auto" }}
-  >
-    <Form.Item name="date-picker" label="DatePicker" {...config}>
-      <DatePicker />
-    </Form.Item>
-    <Form.Item name="date-time-picker" label="DatePicker[showTime]" {...config}>
-      <DatePicker showTime format="YYYY-MM-DD HH:mm:ss" />
-    </Form.Item>
-    <Form.Item name="month-picker" label="MonthPicker" {...config}>
-      <DatePicker picker="month" />
-    </Form.Item>
-    <Form.Item name="range-picker" label="RangePicker" {...rangeConfig}>
-      <RangePicker />
-    </Form.Item>
-    <Form.Item
-      name="range-time-picker"
-      label="RangePicker[showTime]"
-      {...rangeConfig}
-    >
-      <RangePicker showTime format="YYYY-MM-DD HH:mm:ss" />
-    </Form.Item>
-    <Form.Item name="time-picker" label="TimePicker" {...config}>
-      <TimePicker />
-    </Form.Item>
-    <Form.Item label={null}>
-      <Button type="primary" htmlType="submit">
-        Submit
-      </Button>
-    </Form.Item>
-  </Form>
-)
-
-export default TimeRelatedFormExample
